Fix sleeper options in bc-import help
Removed unused `/sleeper` and `/nosleeper` parameters. It was said to overwrite _xml settings_, which, perhaps, indicates importing
a prefab but ignoring it's sleepers? Not implemented as far I can tell from the code, so maybe it was something for the future.

Added `/editmode` and `/sblock`, which means importing sleeper blocks instead of spawning sleepers.